pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["KLan Stovall <kland13@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.1"
django-environ = "^0.8.1"
djangorestframework = "^3.13.1"
whitenoise = "^5.3.0"
django-cors-headers = "^3.11.0"
psycopg2-binary = "^2.9.3"
djangorestframework-simplejwt = "^5.0.0"
docker = "^5.0.3"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
